radius is long

+++ watchFaceParser/elements/gtr2/basicElements/fontRotate.py
@@ -3,6 +3,6 @@ class FontRotate:
     definitions = {
         1: { 'Name': 'X', 'Type': 'long'}, # circle center
         2: { 'Name': 'Y', 'Type': 'long'}, # circle center
-        3: { 'Name': 'Radius', 'Type': 'float'}, # radius from center to middle fontline
+        3: { 'Name': 'Radius', 'Type': 'long'}, # radius from center to middle fontline
         4: { 'Name': 'RotateDirection', 'Type': 'float'}, # 0 - clockwise, 1 - counterclockwise
     }
